requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.7
django-heroku==0.3.1
exceptiongroup==1.1.3
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
multidict==6.0.4
openai==0.28.0
packaging==23.2
psycopg2==2.9.9
pydantic==2.4.2
pydantic-core==2.10.1
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
yarl==1.9.4
